requirements.txt
Flask
Flask_restful
Flask_SQLAlchemy
pymysql
flask_jwt_extended
flask-cors